pyproject.toml
[tool.poetry]
name = "llama3-1"
version = "0.1.0"
description = ""
authors = ["kingpin147 <nomiking0072012@gmail.com>"]
readme = "README.md"
packages = [{include = "llama3"}]

[tool.poetry.dependencies]
python = "^3.12"
groq = "^0.9.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
